feat: 이동 명령어 추가
